Corrects bad mu4e config
Removes the bad config that was a mix of pre-0.9 and post-0.9 compatible mu4e
code. The end result is that mu4e can automatically pick the right email context
nine times out of ten and do the right thing. If mu4e can't determine the
correct account to send from, it will prompt the user.